refactor(ad-cards): rename SortableAdCards to AdCards
